feat(fixtures): add P0 replay harness
